(src/include) Adjusting code
- Removing references to vroom
- Fixing filenames on the license

+++ src/optimize/optimize.c
@@ -163,7 +163,7 @@ _vrp_optimize(PG_FUNCTION_ARGS) {
     }
 
 
-    values[0] = Int32GetDatum(funcctx->call_cntr + 1);
+    values[0] = Int32GetDatum((int32_t) funcctx->call_cntr + 1);
     values[1] = Int64GetDatum(result_tuples[call_cntr].vehicle_id);
     values[2] = Int64GetDatum(result_tuples[call_cntr].order_id);
